added mini server and reorganized
