+++ client/src/components/RelatedProducts/RelatedProducts.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+var RelatedProducts = () => {
+  return <div>RelatedProducts</div>;
+}
+
+export default RelatedProducts;
