+++ client/src/components/Reviews/Reviews.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+var Reviews = () => {
+  return <div>Reviews</div>;
+}
+
+export default Reviews;
