+++ server.js
@@ -1,0 +1,11 @@
+const path = require("path")
+const express = require("express"); // npm installed
+
+const app = express();
+
+app.use(express.static(path.join(__dirname, "/client/dist")));
+// other configuration...
+
+app.listen(3000, () => {
+  console.log('Listening on 3000!');
+});
