+++ webpack.config.js
@@ -2,9 +2,9 @@ const path = require("path");
 
 module.exports = {
   mode: "development",
-  entry: "./src/index.jsx",
+  entry: "./client/src/index.jsx",
   output: {
-    path: path.join(__dirname, 'public'),
+    path: path.join(__dirname, './client/dist'),
     filename: "bundle.js"
   },
   module: {
@@ -27,7 +27,7 @@ module.exports = {
   devServer: {
     // [static] config for how what to serve
     static: {
-      directory: path.join(__dirname, 'public'),
+      directory: path.join(__dirname, '/client/dist'),
     },
     compress: true,
     // [port] what port on our local machine to run the dev server
